requirements.txt
django ~= 3.1.7
Pillow ~= 8.1.2
unidecode ~= 1.2.0
GeoIP2 ~= 4.1.0
djangorestframework ~= 3.12.2
django-cors-headers ~= 3.7.0